+++ shared/src/commonMain/kotlin/app/duss/easyproject/presentation/ui/project/list/store/ProjectListStore.kt
@@ -6,14 +6,14 @@ import com.arkivanov.mvikotlin.core.store.Store
 interface ProjectListStore: Store<ProjectListStore.Intent, ProjectListStore.State, Nothing> {
 
     sealed class Intent {
-        data class LoadProjectListByPage(val page: Long): Intent()
+        data class LoadProjectListByPage(val page: Int): Intent()
         data class UpdateSearchValue(val searchValue: String): Intent()
         data object AddNew: Intent()
         data class Details(val id: Long): Intent()
     }
 
     data class State(
-        var page: Long = 0,
+        var page: Int = 0,
         var projectId: Long? = null,
         val isLoading: Boolean = false,
         val isLastPageLoaded: Boolean = false,
